backend/requirements.txt
altair
annotated-types
anyio
asgiref
attrs
beautifulsoup4
blinker
cachetools
certifi
cffi
charset-normalizer
click
colorama
comtypes
cryptography
deep-translator
distlib
distro
dj-database-url
django
django-environ
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
faiss-cpu
fastapi
filelock
fsspec
gitdb
gitpython
greenlet
gunicorn
h11
httpcore
httpx
huggingface-hub
idna
jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-core
langchain-text-splitters
langsmith
language-tool-python
lxml
markupsafe
narwhals
numpy
openai
orjson
packaging
pandas
pdf2image
pdfminer-six
pdfplumber
pillow
platformdirs
plyer
protobuf
psutil
psycopg
psycopg2-binary
pyarrow
pyaudio
pycparser
pydantic
pydantic-core
pydeck
pyjwt
pypdf2
pypdfium2
pypiwin32
pytesseract
python-dateutil
python-docx
python-dotenv
python-multipart
pyttsx3
pytz
pywin32
pyyaml
referencing
regex
requests
requests-toolbelt
rpds-py
safetensors
scikit-learn
scipy
setuptools
six
smmap
sniffio
soupsieve
speechrecognition
sqlalchemy
sqlparse
starlette
streamlit
tenacity
threadpoolctl
tokenizers
toml
tornado
tqdm
transformers
ttkbootstrap
typing-extensions
typing-inspection
tzdata
urllib3
uvicorn
virtualenv
virtualenvwrapper-win
watchdog
zstandard
nltk